--- a/test/fixtures/alignment_configs.yml
+++ b/test/fixtures/alignment_configs.yml
@@ -1,2 +1,3 @@
 one:
   name: 'ac1'
+  lineage_version: 1