- Tested locally on a couple of samples
- Replace the logic of "date between started_at and ended_at" to be "alignment_config.lineage_version between version_start and version_end" in the lineage calls
- taxon_search_list is meant to be called manually / should be updated later. Or we can add column defaults to version_start/version_end as well.