@cdebourcy 2018-11-29
It would be great if you could add some performance benchmarks, so we can evaluate and be aware of the improvement!