@jshoe 2018-11-29
@cdebourcy 

# Benchmarks results
- Ran some speed tests locally. 
- Benchmark: Load report_info for all the samples in the range
- Results: Overall about 20-30% speedup with this switch

# Samples 900-950 (lots of skipped ones here)
OLD: 13.2179527 sec
NEW: 7.9781604 sec

# Samples 600-650
OLD: 76.796431 sec
NEW: 62.6806688 sec

# Samples 500-550
OLD: 48.5035036 sec
NEW: 34.7451729 sec

- Mixed up the orders of the tests to avoid caching issues
- Comparison code: https://github.com/chanzuckerberg/idseq-web/compare/jsheu/lineage-bench
- 
```
start_time = Time.now
(500..550).each do |i|
  puts "LOADING: " + i.to_s
  SamplesController.new.report_info(i)
end
end_time = Time.now
duration = end_time - start_time
puts "DONE"
puts "Time elapsed: #{duration}"
```